Task #3651: Merged task branch for issue #3719 into this branch.

+++ branches/CEP-Pipeline-Task3651/MAC/APL/CR_Protocol/include/APL/CR_Protocol/CRstopVector.h
@@ -43,9 +43,9 @@ public:
 
 	//@{
 	// --- marshalling methods --- 
-	unsigned int getSize();
-	unsigned int pack  (void* buffer);
-	unsigned int unpack(void *buffer);
+	size_t getSize();
+	size_t pack  (void* buffer);
+	size_t unpack(void *buffer);
 	//@}
 
 	// --- datamembers ---
